requirements.txt
# EnMapper - Core Dependencies with Pinned Versions
# Phase 0: Up-Front Decisions & Contracts

# LLM Framework & Orchestration
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.7
langchain-openai==0.0.2.post1
langchain-anthropic==0.1.1
langgraph==0.0.20
langsmith==0.0.70
crewai==0.22.5

# Data Processing & Analytics
polars==0.20.2
pyarrow==15.0.0
pandas==2.1.4
numpy==1.26.2

# Database & Storage
sqlalchemy==2.0.25
sqlglot==20.11.0
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
redis-py==5.0.1

# Data Validation & Quality
great-expectations==0.18.8
pandera==0.17.2
pydantic==2.5.3

# Units & Measurement
pint==0.23

# API & Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
starlette==0.32.0.post1

# Observability & Monitoring
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utilities
rich==13.7.0
typer==0.9.0
jinja2==3.1.2
